+++ src/shared/calendar/index.tsx
@@ -1,0 +1,4 @@
+export { default as Calendar } from './calendar-provider'
+export { default as CalendarHeader } from './calendar-header'
+export { default as CalendarWeekdays } from './weekday-header'
+export { default as CalendarCells } from './calendar-cells'
